Moved router/admin/util.go to secret/secret.go
